Fix possible loss of data warning in RGBController_MSIMysticLight64.cpp

+++ Controllers/MSIMysticLightController/MSIMysticLight64Controller/RGBController_MSIMysticLight64.cpp
@@ -85,7 +85,7 @@ void RGBController_MSIMysticLight64::DeviceUpdateLEDs()
     unsigned int num_colors = 0;
     if(Mode.flags & MODE_FLAG_HAS_MODE_SPECIFIC_COLOR)
     {
-        num_colors = Mode.colors.size();
+        num_colors = (unsigned int)Mode.colors.size();
         for(unsigned int i = 0; i < num_colors; i++)
         {
             led_colors[i].R = RGBGetRValue(Mode.colors[i]);
